bilibili/requirements.txt
ChatterBot
bilibili-api-python
edge-tts